improve tool with proper cli interface
