+++ lib/src/cli/output.dart
