+++ lib/src/cli/schema.dart
