+++ pubspec.yaml
@@ -6,6 +6,7 @@ issue_tracker: https://github.com/schultek/stormberry/issues
 
 dependencies:
   analyzer: ^3.3.1
+  args: ^2.3.1
   build: ^2.2.1
   collection: ^1.16.0
   crypto: ^3.0.1
